HelpComponent: Display different help icons depending on controller
Before this commit, all buttons shown in the help component would use
the SNES layout, which wouldn't match when having plugged in an Xbox
or Sony Playstation controller.

If there are different kinds of controllers plugged in, the input
mapping shown will be of the last used or inserted. The controller
identification is done via the SDL gamecontroller API.

The intention here is to help solving #442, or at least most of it.